ln-simln-jamming/refactor: rename sink_interceptor to attack_interceptor
In preparation for pulling this attack out so that the simulator can
be generic across attacks, rename sink_interceptor.
