+++ ln-simln-jamming/src/attack_interceptor.rs
